+++ src/pages/Project/ProjectDefault.tsx
@@ -16,7 +16,6 @@ import { callGetProjectByNameAPI, ProjectResponse } from '@services/adapters/pro
 import { useLocation } from 'react-router-dom';
 import { SlideShowOBJ } from './types';
 import img1 from '@statics/images/correlation/correlation1.png';
-import img2 from '@statics/images/correlation/correlation2.png';
 import img3 from '@statics/images/correlation/correlation3.png';
 import img4 from '@statics/images/correlation/correlation4.png';
 
@@ -56,29 +55,22 @@ const ProjectDefault: React.FC<ProjectProps> = (props) => {
 
     const dummyList: SlideShowOBJ[] = [
         {
-            img: img1,
-            title: 'Figure 1.1',
-            description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-        },
-        {
-            img: img2,
-            title: 'Figure 1.2',
-            description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
-        },
-        {
             img: img3,
             title: 'Figure 1.3',
             description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
+            , cardType: 'default'
         },
         {
             img: img4,
             title: 'Figure 1.4',
             description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
+            , cardType: 'default'
         },
         {
             img: img1,
             title: 'Figure 1.5',
             description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
+            , cardType: 'default'
         }
     ]
 
@@ -119,7 +111,6 @@ const ProjectDefault: React.FC<ProjectProps> = (props) => {
             </div>
             <img src={GalleryCog} className="gallery-cog" alt="a design elemnt depicting a set of cogs" />
 
-            {/* todo gallery component */}
             <ProjectGallery displayNumber={2} compTitle={'Gallery'} itemArray={dummyList}/>
 
             <VerticalSpacer height={20} />
