+++ src/pages/Project/types.tsx
@@ -1,5 +1,8 @@
 export interface SlideShowOBJ {
-    img: string;
-    title: string;
+    img?: string;
+    title?: string;
     description: string;
+    cardType: string;
+    name?: string; // only if the card is of type 'testimony'
+    position?: string; // only if the card is of type 'testimony'
 }
